week08/explore_phila_data/extract_data/requirements.txt
functions_framework==3.5.0
google-cloud-storage==2.16.0
python-dotenv==1.0.1
Requests==2.31.0
